adding viewmodel
